added list comprehensions

+++ compiler/lexer/lexer.h
@@ -21,12 +21,12 @@ one of the delimiters ), ], or }
 #define STR_DELIM '\''
 
 typedef struct {
-    FILE *file;
+    FILE *file;     // OWN
     char c;
     Token type;
-    char *value;
-    int64_t val_len;
-    int line;
+    char *value;    // NOT OWN
+    size_t val_len;
+    int64_t line;
 } Lexer;
 
 static Token YASLToken_FourChars(char c1, char c2, char c3, char c4);
